Revert "refactor(path): separate workspace validation from path normalization (#205)"
This reverts commit 71d9db730af0f60e81fd5f9c8e6e12def57345b3.
